+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Bug Fixes
 
 * [UCX-3770] Fix JS exception when selecting from a message popup
+* Trim a message before saving it
 
 ## 0.2.3 (2018-02-02)